% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.trio.R
\name{analyse.trio.set}
\alias{analyse.trio.set}
\title{Infer the causal networks for a set trios}
\usage{
analyse.trio.set(
  trio.set,
  nb.trios = NROW(trio.set),
  Qlabels = NULL,
  alpha,
  FDRcontrol,
  fdr,
  lambda,
  lambda.step,
  pi0.meth = "bootstrap",
  data,
  confounders,
  n_v,
  n_t,
  use.perm,
  gamma,
  is.CNA,
  nperms,
  verbose,
  cl = NULL,
  chunk.size = NULL
)
}
\arguments{
\item{trio.set}{numeric matrix with three column, each row giving the numbers
of the columns of \code{data} defining a target trio.}

\item{nb.trios}{number of trios to analyse. If \code{nb.trios} is less than
the number of rows of \code{trio.set}, only the first \code{nb.trios} trios are
analysed.}

\item{Qlabels}{numeric vector of column numbers indicating the columns of \code{data}
that correspond to common children or intermediate variables in the general network
from which trios are formed.}

\item{alpha}{significance level for individual tests, i.e. the rejection threshold
for each Wald tests. Passed to \link[MRGN]{infer.trio}}

\item{FDRcontrol}{character, method for p-value adjustment for multiple
comparisons. Either \code{'qvalue'}, or any method accepted by \link[stats]{p.adjust}.}

\item{fdr, lambda, pi0.meth}{arguments passed to \link[MRGN]{adjust.q}.}

\item{lambda.step}{used to define the default value of \code{lambda} when the latter is \code{NULL}.}

\item{data}{\code{data.frame} object where columns indexed in \code{trio.set},
\code{Qlabels}, \code{confounders} are taken from for trio analysis.
The first \code{n_t} columns are \code{V}-nodes (variants), the next \code{n_t}
columns are \code{T}-nodes (genes), and the remaining columns are
\code{Q}-nodes (common children and intermediate variables), and
\code{U}-nodes (confounders).}

\item{confounders}{\code{n_t}-list of column numbers indexing confounders for each \code{T}-node in \code{data}.}

\item{n_v, n_t}{integers, numbers of respectively \code{T}-nodes and \code{V}-nodes
in \code{data}.}

\item{use.perm, gamma, is.CNA, nperms, verbose}{arguments passed to \link[MRGN]{infer.trio}.}

\item{cl, chunk.size}{optional arguments for parallel computing, passed to
\link[parallel]{parLapply} (when supplied).}
}
\description{
This function takes a set of trios including each a genetic variant,
phenotypes, and optionally confounding variables, and performs the MRGN inference.
This a wrapper for the function \link[MRGN]{infer.trio} of the package \code{MRGN}.
\code{analyse.trio.set} returns a \code{data.frame} object with one row for each input trio.
}
