% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.conf.selection.R
\name{assess.conf.selection}
\alias{assess.conf.selection}
\title{Performance of confounding variable selection}
\usage{
assess.conf.selection(
  conf.sets,
  adjacency,
  n_v = length(conf.sets$WZconfounders),
  n_t = length(conf.sets$confounders),
  n_w = 0,
  n_z = 0,
  n_u = NCOL(adjacency) - n_v - n_t - n_w - n_z
)
}
\arguments{
\item{conf.sets}{an object of class \code{conf.sets} as returned by \link{get.conf.sets}.}

\item{adjacency}{numeric, adjacency matrix of the true network, i.e., a binary square matrix
with zeros on the main diagonal. The columns and rows of \code{adjacency} must be such
that the first \code{n_v} correspond to genetic variants (\code{V}-nodes),
the next \code{n_t} to target phenotypes/expressions values (\code{T}-nodes),
the next \code{n_w} to intermediate variables (\code{W}-nodes),
the next \code{n_z} to common children (\code{Z}-nodes), and
the last \code{n_u} to confounders (\code{U}-nodes).}

\item{n_v, n_t, n_w, n_z, n_u}{integers, the numbers
\code{n_v} of \code{V}-nodes,
\code{n_t} of \code{T}-nodes,
\code{n_w} of \code{W}-nodes,
\code{n_z} of \code{Z}-node, and
\code{n_u} of \code{U}-nodes in the true network.
Note that the default values inferred from \code{conf.sets} assume
\code{n_w = n_z = 0}, i.e. there is no common child or intermediate variable
in the true network.}
}
\value{
a named list with the following elements:
\describe{
\item{\code{recall}}{a 6-vector giving the recalls for the pool of
\code{W,Z}-nodes, and the pool of \code{U}-nodes, and the average recalls
for \code{V}-nodes, \code{T}-nodes, \code{Q}-nodes (mixture of \code{W} and
\code{Z}-nodes), and \code{U}-nodes. The raw recalls the averages are computed
from are in the returned element \code{raw}.}
\item{\code{precision}}{a 6-vector giving the precisions for the pool of
\code{W,Z}-nodes, and the pool of \code{U}-nodes, and the average precisions
for \code{V}-nodes, \code{T}-nodes, \code{Q}-nodes, and \code{U}-nodes. The
raw precisions the averages are computed from are in the returned element \code{raw}.}
\item{\code{raw}}{a list with elements \code{V}, \code{T}, \code{Q}, and
\code{U}, each element being a named list of two vectors (
\code{recall}: \code{n_t}-vector giving the recall of a particular group of
nodes (e.g. \code{V}, \code{T}, \code{Q}, \code{U}) for each focal \code{T}-node,
\code{precision}: \code{n_t}-vector giving the precisions of a particular
group of nodes for each focal \code{T}-node).}
}
}
\description{
Compute recall and precision for confounding variables
selected using \link{get.conf.sets}.
}
\details{
The function evaluates the performance of \link{get.conf.sets} on a single
genomic dataset, given the adjacency matrix of the true network.
}
\examples{
## Load the 'conf.sets' object 'confsetsA11' and the raw data 'networkA11' it was obtained from
library(MRGNgeneral)
data(confsetsA11)
data(networkA11)

## Recall and precision of the selection procedure
Perf <- assess.conf.selection (confsetsA11,
                               adjacency = networkA11$adjacency,
                               n_v = networkA11$dims$n_v,
                               n_t = networkA11$dims$n_t,
                               n_w = networkA11$dims$n_w,
                               n_z = networkA11$dims$n_z,
                               n_u = networkA11$dims$n_u)

Perf$recall
Perf$precision
}
