% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnclass2adj.R
\name{bnclass2adj}
\alias{bnclass2adj}
\title{From \code{bn} class objects to adjacency matrix}
\usage{
bnclass2adj(bnfit, pattern = "G", nodenames = NULL)
}
\description{
\code{bnclass2adj} converts a list of edges of a \code{bn} class object (e.g.
from the package \code{bnlearn}) into an adjacency matrix.
}
