% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confsetsA11.R
\docType{data}
\name{confsetsA11}
\alias{confsetsA11}
\title{An object of class \code{'conf.sets'}}
\format{
A named list with elements \code{confounder}, \code{Vconfounder},
\code{Tconfounder}, \code{Uconfounder}, \code{confounder},
\code{UWZconfounders}, \code{WZconfounders}, \code{UWZindices},
\code{WZindices}, \code{raw}, \code{time}, and \code{call}.
See the section "Value" of \link{get.conf.sets} for a description of these elements.
}
\source{
Generated in the package \code{MRGNgeneral}.
}
\usage{
confsetsA11
}
\description{
An \code{.rda} structure as returned by \link{get.conf.sets}, i.e.
an object of class \code{'conf.sets'}.
}
\examples{
## Load the raw data (networkA11) used to obtain 'confsetsA11'
library(MRGNgeneral)
data(networkA11)
data(confsetsA11)

## Recall and precision of the selection procedure
Perf <- assess.conf.selection (confsetsA11,
                               adjacency = networkA11$adjacency,
                               n_v = networkA11$dims$n_v,
                               n_t = networkA11$dims$n_t,
                               n_w = networkA11$dims$n_w,
                               n_z = networkA11$dims$n_z,
                               n_u = networkA11$dims$n_u)

Perf$recall
Perf$precision

}
\keyword{datasets}
