% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges2adj.R
\name{edges2adj}
\alias{edges2adj}
\title{From edges to adjacency matrix}
\usage{
edges2adj(edges, pattern = NULL, diag = FALSE)
}
\arguments{
\item{edges}{a \code{matrix}, \code{data.frame} or \code{list}.
\code{edges} must be a numeric 2-column matrix or a data.frame if
\code{pattern} is NULL (or missing).}

\item{pattern}{NULL or a \code{character}, indicates the labels of nodes.
For instance, \code{pattern = 'G'} means that nodes in \code{edges} are
labelled \code{Gi} where \code{i} is a positive integer number.}

\item{diag}{logical, should non null values be allowed on the main diagonal
of the adjacency matrix? Defaults to \code{FALSE}.}
}
\value{
an object of class \code{adjacency.matrix}.
}
\description{
A function to convert a list of edges into an adjacency matrix.
}
\details{
The fitting functions in the package \code{bnlearn} return list of
edges. The function \code{edges2adj} is meant to derive the adjacency matrix
corresponding to such a list.
}
\seealso{
\link{is.adjacency.matrix}.
}
