% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate.new.trios.R
\name{enumerate.new.trios}
\alias{enumerate.new.trios}
\title{Form trios involving genetic variants}
\usage{
enumerate.new.trios(
  trio.set,
  new.edges,
  old.Adj,
  Adj,
  q,
  p,
  r = NROW(Adj) - p - q,
  VTT = TRUE,
  Qtrio.set = NULL,
  cl,
  chunk.size = NULL
)
}
\arguments{
\item{trio.set}{\eqn{n \times 3} matrix of \eqn{n} already analysed trios.
Each row of \code{trio.set} is a vector giving the labels of a genetic variant
(\code{V}-node), and two genes (\code{T}-nodes).}

\item{new.edges}{logical \eqn{n} vector, the \code{i}th element indicating the
presence/absence of a new edge in the trio at the \code{i}th row of \code{trio.set}.}

\item{old.Adj}{Adjacency matrix used to form trios in \code{trio.set}.}

\item{Adj}{Updated adjacency matrix after the analysis of trios in \code{trio.set}.}

\item{q, p, r}{Numbers of respectively \code{V,T,Q}-nodes in \code{Adj}. I.e.,
the first \code{q} columns/rows of \code{Adj} correspond to \code{V}-nodes,
the next \code{p} columns/rows correspond to \code{T}-nodes, and the last
\code{r} columns/rows correspond to \code{Q}-nodes.}

\item{VTT}{logical indicating the composition of desired trios. If true (the default),
only trios not involving a \code{Q}-node are formed, i.e. a trio contains one
\code{V}-node and two \code{T}-nodes. Otherwise, only trios involving
\code{Q}-nodes are formed, i.e. a trio contains one \code{V}-node,
one \code{T}-node, and one \code{Q}-node.}

\item{Qtrio.set}{\eqn{m \times 3} matrix of \eqn{m} already analysed trios
with \code{Q}-nodes. Each row of \code{trio.set} is a vector giving the labels
of a genetic variant (\code{V}-node), one genes (\code{T}-node), and one common
child or intermediate (\code{Q}-node). Used when \code{VTT = FALSE} to sort
out any previously analysed trio from the return output.}

\item{cl}{a cluster object, created by one of the packages \code{parallel} and \code{snow}.
Specify \code{cl = NULL} for no parallel computation. Note that this is not the
genuine default. If missing, the registered default cluster (which can be
\code{NULL}, but may not be) is considered.}

\item{chunk.size}{integer, number of tasks per scheduling unit during parallel computation.}
}
\value{
An \eqn{K \times 3} matrix of \eqn{K} trios. Each row gives the
column numbers of variables forming a trio, the first giving the genetic
variant (\code{i}), and the last two giving phenotypes, named 'Tj', 'Tk'
when \code{VTT = TRUE}, and named 'Tj', Qk' when \code{VTT = FALSE}.
If no additional trio can be formed, the returned matrix has zero row.
}
\description{
Given some already analysed trios, enumerate new trios involving
genetic variants. Not a user level function.
}
\seealso{
\link{enumerate.trios} to form trios when none has been yet analysed.
}
