% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate.trios.R
\name{enumerate.trios}
\alias{enumerate.trios}
\title{Form trios involving genetic variants}
\usage{
enumerate.trios(
  i = 1:q,
  Adj,
  p = NROW(Adj) - q,
  q = length(i),
  r = NROW(Adj) - p - q,
  VTT = TRUE,
  cl,
  chunk.size = NULL
)
}
\arguments{
\item{i}{integer vector, position of the target \code{V}-nodes (genetic
variants) in the adjacency matrix \code{Adj}: \code{i} must satisfy
\code{1 ≤ i ≤ q} (see the argument \code{q}). For convenience,
\code{i} is coerced to the integer type using the function \link{as.integer}.
Defaults to \code{i = 1:q} (i.e. all genetic variants).}

\item{Adj}{numeric, an adjacency matrix, i.e., a binary symmetric matrix
(graph object). The first \code{q} columns of \code{Adj} represent
\code{V}-nodes, and the last \code{p} columns of \code{Adj} represent
\code{T}-nodes (expressions/phenotypes). Only the upper triangular part of
\code{Adj} is actually used.}

\item{p, q, r}{integers, respectively: number of \code{T}-nodes (non-instrumental variables),
number of \code{V}-nodes (instrumental variables), and number of \code{Q}-nodes
(confounding variables: intermediate variables and common children).}

\item{VTT}{logical indicating the composition of desired trios. If true (the default),
only trios not involving a \code{Q}-node are formed, i.e. a trio contains one
\code{V}-node and two \code{T}-nodes. Otherwise, only trios involving
\code{Q}-nodes are formed, i.e.  a trio contains one \code{V}-node,
one \code{T}-node, and one \code{Q}-node.}

\item{cl}{a cluster object, created by one of the packages \code{parallel} and \code{snow}.
Specify \code{cl = NULL} for no parallel computation. Note that this is not the
genuine default. If missing, the registered default cluster (which can be
\code{NULL}, but may not be) is considered.}

\item{chunk.size}{integer, number of tasks per scheduling unit during parallel computation.}
}
\value{
A list of matrices, one for each unique element of \code{i}. If the
input adjacency matrix \code{Adj} has non-\code{NULL} colnames (as returned
by \link{colnames}), then the elements of the list are named after the
corresponding columns (from the first \code{q} columns). Otherwise, they are
named as 'Vi' where 'i' are the unique elements of \code{i}.

For each matrix, each row gives the column numbers of variables forming a
trio, the first giving the genetic variant (\code{i}), and the last two
giving phenotypes, named 'Tj', 'Tk' when \code{VTT = TRUE}, and named 'Tj',
'Qk' when \code{VTT = FALSE}. If no trio can be formed for a given genetic
variant, the associated matrix has zero row.
}
\description{
Enumerate all trios involving some target genetic variants based on
an adjacency matrix.
Not a user level function.
}
\details{
When \code{i} has more than one elements, trio enumeration is performed
for each of its unique elements (as returned by \link{unique}).

In a trio, the variant \code{Vi} has an edge with at least one of the two
other variables in the trio. In addition to the variant, any trio has either
two different \code{T}-nodes, or one \code{T}-node and one \code{Q}-node: a
trio with two different \code{Q}-nodes is not allowed.
}
\seealso{
\link{enumerate.new.trios} to enumerate additional trios after updating the
adjacency matrix \code{Adj}.

\link{enumerate.triplets} to enumerate trios involving no genetic
variant (only \code{T} and \code{Q}-nodes).
}
