% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate.triplets.R
\name{enumerate.triplets}
\alias{enumerate.triplets}
\title{List all (update-able) trios involving only phenotypes}
\usage{
enumerate.triplets(
  Adj,
  p,
  q,
  r = NROW(Adj) - p - q,
  TTT = TRUE,
  return.3 = FALSE,
  cl,
  chunk.size = NULL
)
}
\arguments{
\item{Adj}{numeric, a binary matrix indicating adjacency and possibly edge
directions (i.e. a non-directed, or a partially directed graph). All the
\code{p} columns of \code{Adj} represent \code{T}-nodes (phenotypes). As such,
\code{Adj} will in general be an extract of a larger adjacency matrix (last
\code{p} rows and columns in this package). See Section \code{Details} for more
on the interpretation of the values \code{0/1} in \code{Adj} for this function.}

\item{p, q, r}{integers, respectively: number of \code{T}-nodes (non-instrumental variables),
number of \code{V}-nodes (instrumental variables), and number of \code{Q}-nodes
(confounding variables: intermediate variables and common children).}

\item{TTT}{logical indicating the composition of desired trios. If true (the default),
only trios not involving a \code{Q}-node are formed, i.e. a trio contains one
three \code{T}-nodes. Otherwise, only trios involving \code{Q}-nodes are
formed, i.e.  a trio contains two \code{T}-nodes, and one \code{Q}-node.}

\item{return.3}{logical, should triplets of type (3) be returned as an
attribute named \code{triplet.3} (of class \code{matrix}) for the output?
Defaults to \code{FALSE}. See Section \code{Details} for more on type (1)
triplets.}

\item{cl}{a cluster object, created by one of the packages \code{parallel} and \code{snow}.
If \code{NULL}, the registered default cluster is used. Note that the latter can be
\code{NULL} too, in which case, no parallel computation is performed.}

\item{chunk.size}{integer, number of tasks per scheduling unit during parallel computation.}
}
\value{
A list with two elements:

\item{triplets}{an \code{n} by \code{3} matrix. Each row of the output gives
the indices of non-instrumental variables forming a triplet with two edges.
The columns of the matrix are named 'Ti', 'Tj', and 'Tk'. The \code{T}-nodes
are arranged such that "Tj" has two edges. If no update-able triplet can be
formed, the returned matrix has zero row.}

\item{types}{a \code{n}-vector of integers. The \code{i}th element indicates the
type of triplet corresponding to the i^th row the matrix \code{triplets}.}

If \code{return.3 = TRUE}, then the output matrix \code{triplets} has an
attribute named \code{triplet.3} giving triplets of type (3). See Section
\code{Details} for more on type (3) triplets. The attribute \code{triplet.3}
is a \code{matrix}, and it has zero row if no triplet of type (3) is
found.
}
\description{
Enumerate update-able triplets involving only expressions/phenotypes
(no genetic variant) based on an adjacency matrix. Not a user level function.
}
\details{
The binary matrix \code{Adj} represent a non-directed, or a partially
directed graph. For two nodes \code{Ti} and \code{Tj}, the matrix elements
\code{a_ij} and \code{a_ji} define both the presence of an edge and the
direction of the edge if any: an edge is present between nodes \code{Ti}
and \code{Tj} if \code{a_ij + a_ji > 0}. If \code{a_ij + a_ji = 2}, then the
edge is bi-directed (or equivalently, undirected). Otherwise, if \code{a_ij = 1},
then the edge goes from \code{Ti} to \code{Tj}, and if \code{a_ji = 1}, the
edge goes from \code{Tj} to \code{Ti}.

The function only returns trios for which graph models can be further
identified under the extended interpretation of the Principle of Mendelian
Randomization (PMR). For trios of strictly \code{T}-nodes with at
least one undirected edge, two general situations are possible: (A) the trio
has two edges; and (B) the trio has three edges. In situation (B), no further
inference is possible, and such triplets are \emph{not} returned.

In situation (A), we have three possibilities: (1) two undirected edges (\code{T1-T2-T3});
(2) one directed edge, and the parent \code{T}-node has only one edge (\code{T1->T2-T3});
and (3) one directed edge, and the parent \code{T}-node has two edges (\code{T1<-T2-T3}).

In case (1), one of the edges of the triplet can potentially be directed using
a conditional independence test between the two \code{T}-nodes not directly
related by an edge. Here, the ability to direct the second edge depends on
the test result. Such triplets are returned in the outputed matrix.

In case (2), the parent \code{T}-node can be treated as a genetic variant.
The second edge of the triplet can be directed using a conditional independence
test between the parent \code{T}-node and the non directly-related \code{T}-node.
Such triplets are returned in the outputed matrix.

In case (3), no further inference is possible under PMR. In the absence of an
additional genetic variant linked to one or many of the \code{T}-nodes, we
have three different but Markov equivalent graph candidates, i.e., these
graphs share the same set of conditional and marginal independence relations
(see Kvamme and Fu, 2024). Such triplets are NOT returned in the outputed
matrix. They are nevertheless returned as an attribute named \code{triplet.3}
(of class \code{matrix}) for the matrix output if \code{return.3 = TRUE}
(note that the default is \code{return.3 = FALSE}).

In the outputed matrix, the columns are named 'Ti', 'Tj', and 'Tk', and each
row indicates three \code{T}-nodes whose edges can potentially be updated
(i.e. directed). The \code{T}-nodes are arranged such that 'Tj' has two
edges (\code{Ti-Tj-Tk}). This way, any further inference only requires regressing 'Ti' (or 'Tk')
on 'Tj' and 'Tk' (or 'Ti'), and confounding variables, if any. If a triplet
is of type \code{2}, then 'Ti' is the parent node (i.e. 'Ti' points to 'Tj').
}
\seealso{
\link{enumerate.trios}.
}
