% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_adj.mat.R
\name{is.adjacency.matrix}
\alias{is.adjacency.matrix}
\title{Test if an object is an adjacency matrix}
\usage{
is.adjacency.matrix (object)
}
\arguments{
\item{object}{any \code{R} object.}
}
\value{
logical indicating if \code{object} is an adjacency matrix or not.
}
\description{
Function to test if \code{object} is an adjacency matrix or not.
}
\details{
An adjacency matrix is restricted in this context to represent
acyclic graph objects (directed or not). An adjacency matrix is thus required
to be a square matrix with binary elements (\code{0} or \code{1}), and only
zeros on the main diagonal.
}
\examples{


Adj <- matrix(c(0, 1, 1,
                0, 0, 0,
                1, 1, 0),
                nrow = 3, byrow = TRUE)

is.adjacency.matrix (Adj)

is.adjacency.matrix (Adj + diag(1))


}
