% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.conf.sets.R
\name{is.conf.sets}
\alias{is.conf.sets}
\title{Test if an object is a conf.sets object}
\usage{
is.conf.sets(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Logical, \code{TRUE} if \code{x} is a \code{'conf.sets'} class object, \code{FALSE} otherwise
}
\description{
Tests if its argument is a \code{'conf.sets'} class object.
}
\examples{
## Load data
library(MRGNgeneral)
data(confsetsA11)

## Test if 'confsetsA11' is a 'conf.sets' object
is.conf.sets(confsetsA11)

## Test if a list of one element 'a = 0' is a 'conf.sets' object
is.conf.sets(list(a = 0))
}
