% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbind.R
\name{lbind}
\alias{lbind}
\title{Combine list objects}
\usage{
lbind(
  ...,
  FUN = c,
  unique = TRUE,
  keep.element.names = TRUE,
  force = FALSE,
  make.vector.names = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{...}{R lists objects of presumably the same length \code{n}.
Each list should have \code{n} elements, but each element of a given list can have any length.}

\item{FUN}{the \link{function} to be used to combine list elements. Defaults to \link{c}.

All other arguments below are used only if \code{FUN} is missing or
\code{force = TRUE}. Otherwise, they are ignored. These arguments were added
just for the defaults function \link{c}. They do not necessary make sens for
user provided functions \code{FUN}. If it is certain that these arguments
make sense, then setting \code{force = TRUE} allows their use.}

\item{unique}{logical, return as \code{j}th list element, the union of the \code{j}th elements of arguments \code{...}?
Defaults to \code{TRUE}. The alternative is to concatenate all elements, with possibly duplicates returned.}

\item{keep.element.names}{logical, keep the names of the elements elements that are combined?
Only used when \code{unique = TRUE}, in which case setting \code{keep.element.names = TRUE}
will remove duplicates but will keep the name attributes of the list elements, if any.}

\item{force}{logical, should additinal arguments (after \code{force}) be considered even if \code{FUN} is user-specified?}

\item{make.vector.names}{logical, name the elements of the returned list?
If \code{TRUE} (the default), the \code{name} attribute of the elements of the
\code{...} arguments with the largest length is copied to the returned result
(note that this \code{name} attribute can be \code{NULL}, resulting in a
\code{NULL} \code{name} attribute (i.e. no \code{name}!) for the result).}

\item{verbose}{logical, should warnings be issued when the \code{...} arguments
have different lengths? Defaults to \code{TRUE}.}
}
\description{
Take a sequence of list arguments and combine element-wise.
}
\details{
All arguments after \code{...} must be named.

The elements of the \code{...} arguments are combined using \code{FUN}.
For the defaults function \link{c}, if all or some of these elements were
matrices, they will first be turned  into a vector (stacking all elements),
and then combined with the corresponding elements from other arguments in \code{...}.

All arguments after \code{FUN} are used only if \code{FUN} is missing or
\code{force = TRUE}. Otherwise, they are ignored. These arguments were added
just for the defaults function \link{c}. They do not necessary make sense for
user provided functions \code{FUN}. Setting \code{force = TRUE} allows to
use these arguments with a user specified function \code{FUN}.

If the length of any of the \code{...} arguments has a different length,
all missing list elements at a given position are considered \code{NULL},
and the warning "longer argument not a multiple of length of shorter" is
issued when \code{verbose = TRUE}.
}
