% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgninferA11.R
\docType{data}
\name{mrgninferA11}
\alias{mrgninferA11}
\title{An object of class \code{'MRGN'}}
\format{
See the section "Value" of \link{MRGN} for a description.
}
\source{
Generated in the package \code{MRGNgeneral}.
}
\usage{
mrgninferA11
}
\description{
An \code{.rda} structure as returned by \link{MRGN}, i.e.
an object of class \code{'MRGN'}.
}
\examples{
\dontrun{
## Load data (mrgninferA11)
library(MRGNgeneral)
data(mrgninferA11)

## Print a summary
mrgninferA11

}
}
\keyword{datasets}
