% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.cor.R
\name{partial.cor}
\alias{partial.cor}
\alias{marginal2partial.cor}
\title{Calculate partial correlations (matrix)}
\usage{
partial.cor (x,
             ...)

marginal2partial.cor (R,
                      sd = rep(1, p),
                      cov = NULL)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{...}{additional arguments passed to \link{cov}. See Section \code{Details}.}

\item{R}{numeric \code{p}-by-\code{p} symmetric matrix (marginal correlation matrix).}

\item{sd}{numeric vector of positive scalars, standard deviations of the
variables for which \code{R} is the correlation matrix.}

\item{cov}{numeric matrix (marginal covariance matrix). When supplied and not
\code{NULL}, \code{cov} is used whereas \code{R} and \code{sd} are ignored.}
}
\value{
Either \code{partial.cor} or \code{marginal2partial.cor} returns a square
matrix of partial correlations.
}
\description{
\code{partial.cor} calculates the partial correlation matrix
(i.e. the correlation between each pair of variables given all other
variables) starting from a covariance matrix. \code{marginal2partial.cor}
converts a marginal correlation matrix (with the associated vector of
standard deviations) or the corresponding covariance matrix into a partial
correlation matrix.
}
\details{
The arguments \code{...} include 'use' (which can be one of the
strings "everything", "all.obs", "complete.obs", "na.or.complete", or
"pairwise.complete.obs", or an abbreviation of these strings) and 'method'
(which can here be one of "pearson", or "spearman"). Note that these
arguments are not checked before they are passed to \link{cov}.
}
\seealso{
\link[MESS]{pairwise.cor.test}
}
