% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.sets.R
\name{pcorTTsymmetric}
\alias{pcorTTsymmetric}
\title{Compute T-T partial correlation with fallback logic}
\usage{
pcorTTsymmetric(jk, V.pool, Vconfounders, data, n_v, n_samples)
}
\arguments{
\item{jk}{vector of two T-node column indices in data}

\item{V.pool}{vector of V-node column indices}

\item{Vconfounders}{list of selected V-nodes for each T}

\item{data}{the data matrix}

\item{n_v}{number of V-nodes}

\item{n_samples}{sample size}
}
\value{
list with r (correlation), method used, and n_cond (number of conditioning variables)
}
\description{
When sample size is too small for partial correlation with all V-nodes,
this function falls back to using selected V-nodes (union for each pair),
or marginal correlation if selected V-nodes are still too many.
}
