% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.adjacency.matrix.R
\name{plot.adjacency.matrix}
\alias{plot.adjacency.matrix}
\title{Plot adjacency matrices}
\usage{
plot.adjacency.matrix(
  Adj,
  number.of.V.T = NULL,
  mode = c("directed", "undirected", "max", "min", "upper", "lower", "plus"),
  layout = igraph::layout_nicely,
  vertex.color = "green",
  edge.arrow.size = 0.2,
  ...
)
}
\arguments{
\item{Adj}{A square matrix of class \code{adjacency}. This is passed to
\link{is.adjacency.matrix} for checking.}

\item{number.of.V.T}{numeric, the total number of \code{V} and \code{T} nodes
of the graph represented by \code{Adj}. The default (\code{NULL}) is all. When
specified, \code{number.of.V.T} must be less than or equal to \code{NCOL(Adj)}.}

\item{mode}{Character scalar, specifies how \code{igraph} should interpret the
supplied matrix. See \code{graph_from_adjacency_matrix} in the library
\code{igraph} for details.}

\item{layout}{Either a function or a numeric matrix. It specifies how the
vertices will be placed on the plot. See \code{igraph.plotting} in the
library \code{igraph} for details.}

\item{vertex.color, edge.arrow.size, ...}{Additional plotting parameters.
See \code{igraph.plotting} in the library \code{igraph} for the complete list.}
}
\value{
Returns \code{NULL}, invisibly.
}
\description{
This function takes any adjacency matrix representing an acyclic graph
and calls \link{igraph::plot.igraph} to plot the graph.
}
