% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.nre.run.methods.R
\name{reorder.nre.run.methods}
\alias{reorder.nre.run.methods}
\title{Function to test the stability of network inference method
Re-order data columns and re-run network inference methods}
\usage{
\method{reorder}{nre.run.methods}(
  seed = NULL,
  simdata,
  conf.sets,
  Adjacency0,
  alpha,
  FDRcontrol,
  fdr,
  lambda,
  analyse.triplets,
  stringent,
  nb.cl,
  indexT,
  bn.methods = "none",
  whitelist0,
  blacklist0,
  restart.hc,
  verbose
)
}
\description{
Function to test the stability of network inference method
Re-order data columns and re-run network inference methods
}
