% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.methods2_m.R
\name{run.methods2_m}
\alias{run.methods2_m}
\title{MRGN and \code{bnlearn} methods}
\usage{
run.methods2_m(
  simdata = NULL,
  n_t = 100,
  n_v.t = 1,
  conf.num.vec = c(W = 0, Z = 0, U = 200, K = 0, I = 100),
  graph_type = "scale-free",
  degree = 3,
  theta = 0.4,
  b0 = 0,
  b.snp = c(-0.5, 0.5),
  b.med = c(-0.8, 0.8),
  sigma = 0.1,
  neg.freq = 0.5,
  conf.coef.ranges = list(W = c(0.4, 0.5), Z = c(1, 1.5), U = c(0.4, 0.5), K = c(0.4,
    0.5)),
  sample.size = 500,
  seed = NULL,
  conf.sets = NULL,
  blocksize = min(n_v, n_t, 100),
  T.measure = "partial",
  C.measure = "partial",
  selection_FDRcontrol = "qvalue",
  T.FDRcontrol = selection_FDRcontrol,
  C.FDRcontrol = selection_FDRcontrol,
  V.FDRcontrol = selection_FDRcontrol,
  adjust_by = "individual",
  T.adjust_by = adjust_by,
  C.adjust_by = adjust_by,
  V.adjust_by = adjust_by,
  selection_alpha = 0.01,
  selection_fdr = 0.05,
  lambda = 0.05,
  pi0.method = "smoother",
  bn.methods = c("none", "tabu", "hc", "pc.stable", "mmhc", "fast.iamb", "mmpc", "hpc"),
  run.MRGN = TRUE,
  MRGNfit = NULL,
  threshold_v = 0.2,
  threshold_m = 0.05,
  alpha = 0.05,
  FDRcontrol = "bonferroni",
  fdr = 0.05,
  analyse.triplets = TRUE,
  stringent = TRUE,
  restart.hc = 10,
  n.reorder = 0,
  verbose = 2L,
  nb.cl = 4,
  savetopath = FALSE,
  path = "",
  setpath = ""
)
}
\description{
Run MRGN and selected \code{bnlearn} methods on one simulated genomic dataset.
Depends on package \code{bnlearn} which is not imported and loaded with MRGNgeneral.
}
