% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.graph.data.R
\name{sim.graph.data}
\alias{sim.graph.data}
\title{Generate data from a DAG}
\usage{
sim.graph.data(
  number.of.V,
  number.of.T,
  conf.num.vec,
  graph_type = "scale-free",
  method,
  degree = 3,
  connection_prob = 0.05,
  mixed = "None",
  theta = 0.5,
  b0.1 = 0,
  b.snp = 1,
  b.med = 0.8 * b.snp,
  sd.1 = 1,
  neg.freq = 0.5,
  conf.coef.ranges = list(K = c(0.01, 0.1), U = c(0.15, 0.5), W = c(0.15, 0.5), Z = c(1,
    1.5)),
  sample.size
)
}
\description{
Description
}
