% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.adjacency.R
\name{update.adjacency.matrix}
\alias{update.adjacency.matrix}
\title{Update an adjacency matrix}
\usage{
\method{update}{adjacency.matrix}(
  Adj,
  p,
  q,
  r,
  trio.set,
  inferred.models,
  stringent = FALSE,
  add.edges = TRUE,
  solve.conflicts = TRUE,
  method = "naive",
  added.edges = NULL,
  dropped.edges = NULL,
  cl = NULL,
  chunk.size = NULL,
  ...
)
}
\arguments{
\item{Adj}{numeric, an adjacency matrix}

\item{q}{integer scalar, number of genetic variants in \code{Adj},
not required when \code{solve.conflicts = FALSE}.}

\item{trio.set}{numeric, a matrix where each row indicates the column numbers
of variables forming a trio.

A 3-column matrix \code{trio.set} corresponds to trios involving genetic
variant(s), i.e. each row of \code{trio.set} indicates the column number
of a genetic variant (first element), and the column numbers of two
expressions/phenotypes (second and third columns).

A 4-column matrix \code{trio.set} corresponds to trios involving only
expressions/phenotypes, no genetic variant(s), i.e. each row of \code{trio.set}
indicates the column numbers of three T-nodes, and the type of triplet
(last column). See \link{enumerate.triplets} for details on triplet types.}

\item{inferred.models}{a character vector of length the number of rows in
\code{trio.set}. Each elements of \code{inferred.models} must be one of
\code{M0.1}, \code{M0.2}, \code{M1.1}, \code{M1.2}, \code{M2.1}, \code{M2.2},
\code{M3}, \code{M4}, or \code{Other}. See Badsha and Fu (2019) for the
definitions of these model topologies.}

\item{stringent}{logical, should edges absent from trio topologies \code{M1.1},
\code{M1.2}, \code{M2.1}, and \code{M2.2} be considered absent from the graph?}

\item{solve.conflicts}{logical, should a resolution of conflicts be attempted?
If \code{FALSE}, edges are updated sequentially and conflicts are not noticed.
This makes the final inference depends on the order in which trios are analysed.

If \code{TRUE} (the default), all edges are updated once. This put conflicts
into evidence and requires a method to deal with each type of errors.}

\item{method}{character, only used if \code{solve.conflicts = TRUE}.
The method to be used to solve conflicts. The available \code{method} are:

\code{naive}: conflicts about the presence of an edge are solve by including
an edge when at least one trio analysis inferred an edge; conflicts about an
edge direction are solve by letting the edge undirected.
\code{add_methods}: add further method(s).}

\item{added.edges}{character vector indicating all edges (and directions)
previously added into the network. Only used when \code{solve.conflicts = TRUE}.}

\item{dropped.edges}{character vector indicating all edges (and directions)
previously dropped from the network (or inferred as non existent). Only used
when \code{solve.conflicts = TRUE}.}

\item{cl}{a cluster object, created by one of the packages \code{parallel} and \code{snow}.
If \code{NULL}, the registered default cluster is used. Note that the latter can be
\code{NULL} too, in which case, no parallel computation is performed.}

\item{...}{additional arguments passed to or from other methods.
Currently none.}
}
\description{
This function updates the adjacency matrix of a graph based on
trio analysis. It is used to perform \emph{Step} \code{3.3}, \emph{Step} \code{3.6},
\emph{Step} \code{4.3} and \emph{Step} \code{4.6} of the \code{MRGN} algorithm.
This is not a user level function.
}
\details{
\code{update.adjacency.matrix} takes an adjacency matrix, a set
\code{trio.set} of trios involving genetic variant(s) or not, and the underlying
structures under the extended Mendelian Randomization Principle, and update
the adjacency matrix accordingly.

The registered default cluster is found using \code{parallel::getDefaultCluster()}.
}
