% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.initial_skeleton.R
\name{get.initial.skeleton}
\alias{get.initial.skeleton}
\title{#' Build the input graph for MRGN inference}
\usage{
get.initial.skeleton(
  data,
  n_v,
  n_t,
  threshold_v = 0.2,
  threshold_m = 0.05,
  conf.sets = NULL
)
}
\arguments{
\item{data}{Matrix or data frame where rows are individuals/samples and columns
are variables. The first \code{n_v} columns should be variants (V-nodes),
followed by \code{n_t} phenotypes (T-nodes), and then candidate confounding
variables (C-nodes)}

\item{n_v}{Integer, number of genetic variants (V-nodes) in the network}

\item{n_t}{Integer, number of target phenotypes (T-nodes) in the network}

\item{threshold_v}{Numeric, threshold for determining V-T edges based on
absolute correlation. Default is 0.2. Edges with correlation >= threshold_v
are included in the skeleton}

\item{threshold_m}{Numeric, threshold for determining T-T edges based on
absolute correlation (marginal or partial). Default is 0.05. Edges with
correlation >= threshold_m are included in the skeleton}

\item{conf.sets}{an optional object of class \code{conf.sets} as returned by
\link{get.conf.sets}.}
}
\value{
an object of class \code{adjacency.matrix}, i.e. a binary square matrix
with \code{n_v + n_t + n_q} columns, with only zeros on the main diagonal.

The number \code{n_q} of intermediate variables and common children is taken
from the argument \code{conf.sets} as \code{n_q = length(conf.sets$WZindices)}.
If \code{conf.sets = NULL}, then \code{n_q = 0}.
}
\description{
Build an undirected adjacency matrix for a genomic network using a dataset
where rows are individuals and columns are nodes including \code{n_v} variants
(\code{V}-nodes), \code{n_t} phenotypes (\code{T}-nodes), and \code{n_c}
candidate confounding variables (\code{C}-nodes).
}
\details{
This function implements \code{Step 2} of the MRGN general algorithm.
}
\examples{
\dontrun{
#Raw data (networka11) & confounding variable selection result (confsetsa11)
library(MRGNgeneral)
data(networka11)
data(confsetsa11)

#Graph skeleton
Adjacency0 <- get.initial.skeleton (data = networka11$data,
                                    n_v = networka11$dims$n_v,
                                    n_t = networka11$dims$n_t,
                                    threshold_v = 0.2,
                                    threshold_m = 0.05,
                                    conf.sets = confsetsa11)


#Recall and precision of the skeleton
Recall_Precision <- RecallPrecision(as (networka11$adjacency, 'graphNEL'),
                                    as (Adjacency0, 'graphNEL'),
                                    GV = networka11$dims$n_v,
                                    includeGV = TRUE,
                                    edge.presence = 1.0,
                                    edge.direction = 0.5)
}
}
