% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.initial_skeleton.R
\name{get.initial.skeleton}
\alias{get.initial.skeleton}
\title{Build the input graph for MRGN inference}
\usage{
get.initial.skeleton(
  data,
  n_v,
  n_t,
  threshold_v = 0.2,
  threshold_m = 0.05,
  conf.sets = NULL
)
}
\arguments{
\item{conf.sets}{an optional object of class \code{conf.sets} as returned by
\link{get.conf.sets}.}
}
\value{
an object of class \code{adjacency.matrix}, i.e. a binary square matrix
with \code{n_v + n_t + n_q} columns, with only zeros on the main diagonal.

The number \code{n_q} of intermediate variables and common children is taken
from the argument \code{conf.sets} as \code{n_q = length(conf.sets$WZindices)}.
If \code{conf.sets = NULL}, then \code{n_q = 0}.
}
\description{
Build an undirected adjacency matrix for a genomic network using a dataset
where rows are individuals and columns are nodes including \code{n_v} variants
(\code{V}-nodes), \code{n_t} phenotypes (\code{T}-nodes), and \code{n_c}
candidate confounding variables (\code{C}-nodes).
}
\details{
This function implements \code{Step 2} of the MRGN general algorithm.
}
\examples{
#Raw data (networkA11) & confounding variable selection result (confsetsA11)
library(MRGNgeneral)
data(networkA11)
data(confsetsA11)

#Graph skeleton
Adjacency0 <- get.initial.skeleton (data = networkA11$data,
                                    n_v = networkA11$dims$n_v,
                                    n_t = networkA11$dims$n_t,
                                    threshold_v = 0.2,
                                    threshold_m = 0.05,
                                    conf.sets = confsetsA11)


#Recall and precision of the skeleton
Recall_Precision <- RecallPrecision(as (networkA11$adjacency, 'graphNEL'),
                                    as (Adjacency0, 'graphNEL'),
                                    GV = networkA11$dims$n_v,
                                    includeGV = TRUE,
                                    edge.presence = 1.0,
                                    edge.direction = 0.5)

}
