% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.conf.sets.R
\name{is.conf.sets}
\alias{is.conf.sets}
\title{\code{is.conf.sets} tests if an object is of class \code{'conf.sets'}.}
\usage{
is.conf.sets (x)
}
\arguments{
\item{x}{an object of class 'conf.sets'}
}
\value{
\code{is.conf.sets} returns a logical, \code{TRUE} is \code{x} is \code{'conf.sets'} class object.
}
\description{
\code{is.conf.sets} tests if an object is of class \code{'conf.sets'}.
}
\details{
The function \code{is.conf.sets} helps stop computations early if a user specified
\code{'conf.sets'} object is not licit.
}
\examples{
## Load data
library(MRGNgeneral)
data(confsetsA11)

## Test if 'confsetsA11' is a 'conf.sets' object
is.conf.sets (confsetsA11)

## Test if a list of one element 'a = 0' is a 'conf.sets' object
is.conf.sets (list(a = 0))

}
