% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGN.R
\name{MRGN}
\alias{MRGN}
\alias{MRGNgeneral}
\title{Infer a causal network using the MRGN algorithm}
\usage{
MRGNgeneral(
  data,
  scale.data = TRUE,
  Qlabels = NULL,
  p,
  q,
  r = length(Qlabels),
  u = m - p - q - r,
  confounders = list(),
  Adj,
  is.CNA = FALSE,
  alpha = 0.01,
  use.perm = TRUE,
  gamma = 0.05,
  nperms = 10000,
  FDRcontrol = c("bonferroni", "qvalue", "none"),
  fdr = 0.05,
  lambda = NULL,
  lambda.step = 0.05,
  pi0.meth = c("bootstrap", "smoother"),
  stringent = TRUE,
  analyse.triplets = TRUE,
  solve.conflicts = TRUE,
  method = "naive",
  maxiter = NULL,
  parallel = FALSE,
  cl = parallel:::getDefaultCluster(),
  chunk.size = NULL,
  verbose = 0L
)
}
\arguments{
\item{data}{\code{data.frame}}

\item{Qlabels}{labels indicating columns of \code{data} which are
\code{Q}-nodes (common children or intermediate variables).}

\item{confounders}{a list of length \code{p} giving the set of known confounders
for each phenotype, as returned by \link{get.conf.sets} with the option
\code{return.for.trios = FALSE}.}

\item{Adj}{input adjacency matrix (binary)}

\item{stringent}{logical, should edges not appearing in \code{'M1'} and
\code{'M2'} trio structures be dropped during adjacency matrix update?
Passed to \link{update.adjacency.matrix}.}
}
\description{
Infer a causal network/graph with directed and undirected edges from observational data.
}
\seealso{
\link{infer.trio} for inferring edges in small networks of one
genetic variant and only two genes.
}
