% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.graph.data.R
\name{sample.graph.data}
\alias{sample.graph.data}
\title{sample.graph.data}
\usage{
sample.graph.data(
  n_t = 100,
  n_v.t = 1,
  family.n_v = NULL,
  conf.num.vec = c(W = 0, Z = 0, U = 200, K = 0, I = 100),
  graph_type = "scale-free",
  method,
  degree = 3,
  theta = 0.4,
  b0 = 0,
  b.snp = c(-0.5, 0.5),
  b.med = c(-0.8, 0.8),
  sigma = 0.1,
  neg.freq = 0.5,
  conf.coef.ranges = list(W = c(0.15, 0.5), Z = c(1, 1.5), U = c(0.15, 0.5), K = c(0.15,
    0.5)),
  scale = FALSE,
  sample.size,
  seed = NULL
)
}
\arguments{
\item{seed}{scalar integer, seed for reproducibility of random number generation.}
}
\description{
Different from \code{sim.graph.data} in the way \code{V}-nodes are
included in the network skeleton: here there are just added (not simulated).
The number of \code{V}-nodes is specified for each \code{T}-node, or sampled
from a count distribution. Each \code{V}-node is only related to only one \code{T}-node.
}
