% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.graph.data.R
\name{sample.graph.data}
\alias{sample.graph.data}
\title{sample.graph.data}
\usage{
sample.graph.data(
  number.of.T,
  number.of.V.T = 1,
  family.number.of.V.T = NULL,
  conf.num.vec = rep(0, 4),
  graph_type = "scale-free",
  method,
  degree = 3,
  connection_prob = 0.05,
  mixed = "None",
  theta = 0.5,
  b0.1 = 0,
  b.snp = 1,
  b.med = 0.8 * b.snp,
  sd.1 = 1,
  neg.freq = 0.5,
  conf.coef.ranges = list(W = c(0.15, 0.5), Z = c(1, 1.5), U = c(0.15, 0.5), K = c(0.01,
    0.1)),
  scale = FALSE,
  sample.size
)
}
\description{
Different from \code{sim.graph.data} in the way \code{V}-nodes are
included in the network skeleton: here there are just added (not simulated).
The number of \code{V}-nodes is specified for each \code{T}-node, or sampled
from a count distribution. Each \code{V}-node is only related to only one \code{T}-node.
}
