% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.graph.data.R
\name{sim.graph.data}
\alias{sim.graph.data}
\title{Generate data from a DAG}
\usage{
sim.graph.data(
  number.of.V,
  number.of.T,
  conf.num.vec,
  graph_type = "scale-free",
  method,
  degree = 3,
  connection_prob = 0.05,
  mixed = "None",
  theta = 0.5,
  b0.1 = 0,
  b.snp = 1,
  b.med = 0.8 * b.snp,
  sd.1 = 1,
  neg.freq = 0.5,
  conf.coef.ranges = list(K = c(0.01, 0.1), U = c(0.15, 0.5), W = c(0.15, 0.5), Z = c(1,
    1.5)),
  sample.size
)
}
\arguments{
\item{number.of.V}{Integer, number of genetic variants (V-nodes)}

\item{number.of.T}{Integer, number of phenotypes (T-nodes)}

\item{conf.num.vec}{Named numeric vector of confounder counts for types K, U, W, Z}

\item{graph_type}{Character, type of graph structure. One of "scale-free", "small-world", or "random graph"}

\item{method}{Character, method for generating the random DAG passed to \code{pcalg::randDAG}}

\item{degree}{Integer, average node degree}

\item{connection_prob}{Numeric, connection probability (not currently used)}

\item{mixed}{Character, mixing type (not currently used)}

\item{theta}{Numeric, minor allele frequency for V-nodes}

\item{b0.1}{Numeric, baseline value}

\item{b.snp}{Numeric, effect size for SNPs}

\item{b.med}{Numeric, effect size for mediators}

\item{sd.1}{Numeric, standard deviation for noise}

\item{neg.freq}{Numeric in (0, 1), frequency of negative effects}

\item{conf.coef.ranges}{Named list of coefficient ranges for confounder types}

\item{sample.size}{Integer, number of samples to generate}
}
\value{
A list containing simulated data, dimensions, and graph structure
}
\description{
Generate simulated data from a directed acyclic graph (DAG) with
genetic variants, phenotypes, and confounding variables.
}
